Completed all tasks: Theming, API integration, state management, routing, UI implementation, deployment, and CI/CD setup
